Add Delete method for the postgres driver in asset-syncer (#1328)
